# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reparto API",
    "version": "0.2.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/pines/importaciones": {
      "post": {
        "summary": "Proponer pines (quedan pendientes de revisión; nada cambia solo)",
        "tags": [
          "pines"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pines": {
                    "minItems": 1,
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "rut": {
                          "type": "string",
                          "maxLength": 20
                        },
                        "direccion": {
                          "type": "string",
                          "maxLength": 600
                        },
                        "lat": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "string",
                              "maxLength": 40
                            }
                          ]
                        },
                        "lng": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "string",
                              "maxLength": 40
                            }
                          ]
                        }
                      }
                    }
                  }
                },
                "required": [
                  "pines"
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recibidas": {
                      "type": "number"
                    },
                    "pendientes": {
                      "type": "number"
                    },
                    "sinLocal": {
                      "type": "number"
                    },
                    "errores": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fila": {
                            "type": "number"
                          },
                          "errores": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "codigo": {
                                  "type": "string"
                                },
                                "mensaje": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "codigo",
                                "mensaje"
                              ],
                              "additionalProperties": false
                            }
                          }
                        },
                        "required": [
                          "fila",
                          "errores"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "recibidas",
                    "pendientes",
                    "sinLocal",
                    "errores"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/pines/propuestas": {
      "get": {
        "summary": "Propuestas de pin por estado (primero las que más se alejan del pin actual)",
        "tags": [
          "pines"
        ],
        "parameters": [
          {
            "schema": {
              "default": "pendiente",
              "type": "string",
              "enum": [
                "pendiente",
                "aceptada",
                "rechazada",
                "sin_local"
              ]
            },
            "in": "query",
            "name": "estado",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            },
            "in": "query",
            "name": "limite",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "propuestas": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "localId": {
                            "type": "string"
                          },
                          "rut": {
                            "type": "string"
                          },
                          "direccion": {
                            "type": "string"
                          },
                          "lat": {
                            "type": "number"
                          },
                          "lng": {
                            "type": "number"
                          },
                          "distanciaActualM": {
                            "type": "number"
                          },
                          "estado": {
                            "type": "string",
                            "enum": [
                              "pendiente",
                              "aceptada",
                              "rechazada",
                              "sin_local"
                            ]
                          },
                          "proponenteId": {
                            "type": "string"
                          },
                          "creadaEn": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "razonSocial": {
                            "type": "string"
                          },
                          "comuna": {
                            "type": "string"
                          },
                          "pinActual": {
                            "type": "object",
                            "properties": {
                              "lat": {
                                "type": "number"
                              },
                              "lng": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "lat",
                              "lng"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "required": [
                          "id",
                          "direccion",
                          "lat",
                          "lng",
                          "estado",
                          "proponenteId",
                          "creadaEn"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "propuestas"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/pines/propuestas/{id}/resolver": {
      "post": {
        "summary": "Aceptar o rechazar una propuesta",
        "tags": [
          "pines"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accion": {
                    "type": "string",
                    "enum": [
                      "aceptar",
                      "rechazar"
                    ]
                  }
                },
                "required": [
                  "accion"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "codigo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    },
                    "detalle": {}
                  },
                  "required": [
                    "codigo",
                    "mensaje"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}